+++ tests/formatters/test_json.py
@@ -1,8 +1,7 @@
 from unittest import TestCase
 
-from xrlint.formatter import FormatterContext
 from xrlint.formatters.json import Json
-from .helpers import get_test_results
+from .helpers import get_test_results, get_context
 
 
 class JsonTest(TestCase):
@@ -10,7 +9,7 @@ class JsonTest(TestCase):
         results = get_test_results()
         formatter = Json()
         text = formatter.format(
-            context=FormatterContext(),
+            context=get_context(),
             results=results,
         )
         self.assertIn('"results": [', text)
@@ -19,7 +18,7 @@ class JsonTest(TestCase):
         results = get_test_results()
         formatter = Json(with_meta=True)
         text = formatter.format(
-            context=FormatterContext(),
+            context=get_context(),
             results=results,
         )
         self.assertIn('"results": [', text)