+++ tests/formatters/test_markdown.py
@@ -2,9 +2,8 @@ from unittest import TestCase
 
 import pytest
 
-from xrlint.formatter import FormatterContext
 from xrlint.formatters.markdown import Markdown
-from .helpers import get_test_results
+from .helpers import get_test_results, get_context
 
 
 class MarkdownTest(TestCase):
@@ -13,6 +12,6 @@ class MarkdownTest(TestCase):
         formatter = Markdown()
         with pytest.raises(NotImplementedError):
             formatter.format(
-                context=FormatterContext(),
+                context=get_context(),
                 results=get_test_results(),
             )