+++ tests/formatters/test_simple.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
+from tests.formatters.helpers import get_context
 from xrlint.config import Config
-from xrlint.formatter import FormatterContext
 from xrlint.formatters.simple import Simple
 from xrlint.result import Message
 from xrlint.result import Result
@@ -23,7 +23,7 @@ class SimpleTest(TestCase):
     def test_no_color(self):
         formatter = Simple(styled=False)
         text = formatter.format(
-            context=FormatterContext(),
+            context=get_context(),
             results=self.results,
         )
         self.assert_output_ok(text)
@@ -32,7 +32,7 @@ class SimpleTest(TestCase):
     def test_color(self):
         formatter = Simple(styled=True)
         text = formatter.format(
-            context=FormatterContext(),
+            context=get_context(),
             results=self.results,
         )
         self.assert_output_ok(text)